requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
chromadb>=0.4.22
pandas>=2.0.0
python-dotenv>=1.0.0
